First Official Edit
Changed the health bar to a more fantasy themed health bar, removed the
backgrounds on some things, such as the Flying Guillotine, the ammo
indicator, and the health indicator. Also applied my font set, and set
up the base for further edits. I moved the ammo and the health bar to a
more viewable location. The item effect meters, in exception of the
cleaver, are there simply for testing use, I didn't actually edit these
files.